Changes/conflict has been merged with main

+++ lib/features/admin/screens/add_products_screen.dart
@@ -59,8 +59,8 @@ class _AddProductScreenState extends State<AddProductScreen> {
         context: context,
         name: productNameController.text,
         description: descriptionController.text,
-        price: double.parse(priceController.text),
-        quantity: double.parse(quantityController.text),
+        price: int.parse(priceController.text),
+        quantity: int.parse(quantityController.text),
         category: category,
         images: images,
       );
