+++ lib/features/admin/services/admin_services.dart
@@ -19,8 +19,8 @@ class AdminServices {
     required BuildContext context,
     required String name,
     required String description,
-    required double price,
-    required double quantity,
+    required int price,
+    required int quantity,
     required String category,
     required List<File> images,
   }) async {
